Turns now work properly
Dicesets are easy, but maintaining immutability in turns means that I'll
have to be creating new turns on the regular, because they refer to a
DiceSet.

I want to do this to be able to maintain state as I go.  Eventually
turns will refer to a list of Turns, grouped into a list of Players
owned by a game.